Add itone and related data to status message packet
This feature is useful for 'digital transceivers' as they NO longer need
to rely on AFP-FSK technique (which introduces a non-zero error into the
transmissions).

While this approach is not as transparent as the AFP-FSK technique, it
eliminates the non-zero error associated with AFP-FSK.

As an alternate approach, the 'digital transceivers' can use ft8_lib to
derive the tones from the 'm_currentMessage' data (which is already a
part of the status message packet). However, ft8_lib is currently unable
to handle all the possible types of input messages.